requirements.txt
# SmartFormsGPT - Python Dependencies

# Core Framework
streamlit==1.29.0

# AI/ML Libraries
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.26.2
pydantic==2.5.3

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging
loguru==0.7.2
